Move note type lookup to notes utils
